tests/submission/Apple/dataset/dataset-set.html

<!DOCTYPE html>
<html>
  <head>
    <title>Dataset - Set</title>
    <script src="../../../resources/testharness.js"></script>
    <script src="../../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <h1>Dataset - Set</h1>
    <div id="log"></div>
    <script>
      function testSet(prop, expected)
      {
        var d = document.createElement("div");
        d.dataset[prop] = "value";
        return d.getAttribute(expected) == "value";
      }

      test(function() { assert_true(testSet('foo', 'data-foo')); });
      test(function() { assert_true(testSet('fooBar', 'data-foo-bar')); });
      test(function() { assert_true(testSet('-', 'data--')); });
      test(function() { assert_true(testSet('Foo', 'data--foo')); });
      test(function() { assert_true(testSet('-Foo', 'data---foo')); });
      test(function() { assert_true(testSet('', 'data-')); });
      test(function() { assert_true(testSet('\xE0', 'data-\xE0')); });

      test(function() { assert_throws('SYNTAX_ERR', function() { testSet('-foo', 'dummy') }); });
      test(function() { assert_throws('INVALID_CHARACTER_ERR', function() { testSet('foo\x20', 'dummy') }); });
      test(function() { assert_throws('INVALID_CHARACTER_ERR', function() { testSet('foo\uF900', 'dummy') }); });
    </script>
  </body>
</html>
